# gyeom/week8/dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserMissionAddResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "유저 미션 ID"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"description": "유저 ID"
					},
					"missionId": {
						"type": "number",
						"format": "double",
						"description": "미션 ID"
					},
					"status": {
						"type": "string",
						"description": "미션 상태 (CHALLENGING / DONE)"
					}
				},
				"required": [
					"id",
					"userId",
					"missionId",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserMissionAddResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserMissionAddResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserMissionAddApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserMissionAddResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserMissionAddRequest": {
				"properties": {
					"missionId": {
						"type": "number",
						"format": "double",
						"description": "도전할 미션 ID"
					}
				},
				"required": [
					"missionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserMissionListResponse": {
				"properties": {
					"userMissions": {
						"items": {
							"$ref": "#/components/schemas/UserMissionAddResponse"
						},
						"type": "array",
						"description": "유저 미션 목록"
					}
				},
				"required": [
					"userMissions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserMissionListResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserMissionListResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserMissionListApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserMissionListResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UserMissionCompleteResponse": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double",
						"description": "업데이트된 미션 수"
					}
				},
				"required": [
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserMissionCompleteResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserMissionCompleteResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserMissionCompleteApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserMissionCompleteResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StoreAddResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "가게 ID"
					},
					"name": {
						"type": "string",
						"description": "가게 이름"
					},
					"regionId": {
						"type": "number",
						"format": "double",
						"description": "지역 ID"
					}
				},
				"required": [
					"id",
					"name",
					"regionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_StoreAddResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/StoreAddResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StoreAddApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/StoreAddResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StoreAddRequest": {
				"properties": {
					"name": {
						"type": "string",
						"description": "가게 이름"
					},
					"foodCategoryId": {
						"type": "number",
						"format": "double",
						"description": "음식 카테고리 ID"
					},
					"regionId": {
						"type": "number",
						"format": "double",
						"description": "지역 ID"
					},
					"address": {
						"type": "string",
						"description": "가게 주소"
					},
					"latitude": {
						"type": "number",
						"format": "double",
						"description": "위도"
					},
					"longitude": {
						"type": "number",
						"format": "double",
						"description": "경도"
					}
				},
				"required": [
					"name",
					"foodCategoryId",
					"regionId",
					"address",
					"latitude",
					"longitude"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"MissionAddResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "미션 ID"
					},
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "가게 ID"
					},
					"content": {
						"type": "string",
						"description": "미션 내용"
					},
					"rewardPoint": {
						"type": "number",
						"format": "double",
						"description": "리워드 포인트"
					},
					"deadline": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"description": "마감 기한"
					}
				},
				"required": [
					"id",
					"storeId",
					"content",
					"rewardPoint",
					"deadline"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_MissionAddResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/MissionAddResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionAddApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/MissionAddResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionAddRequest": {
				"properties": {
					"content": {
						"type": "string",
						"description": "미션 내용"
					},
					"rewardPoint": {
						"type": "number",
						"format": "double",
						"description": "미션 달성 시 지급되는 포인트"
					},
					"deadline": {
						"type": "number",
						"format": "double",
						"description": "미션 마감 기한 (선택)"
					}
				},
				"required": [
					"content",
					"rewardPoint"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionListResponse": {
				"properties": {
					"missions": {
						"items": {
							"$ref": "#/components/schemas/MissionAddResponse"
						},
						"type": "array",
						"description": "미션 목록"
					}
				},
				"required": [
					"missions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_MissionListResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/MissionListResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionListApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/MissionListResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "playground-umc-10th-nodejs",
		"version": "1.0.0",
		"description": "UMC 테스트 프로젝트",
		"contact": {}
	},
	"paths": {
		"/users/{userId}/missions": {
			"post": {
				"operationId": "HandleAddUserMission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserMissionAddResponse_"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 사용자 또는 미션입니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 404,
											"message": "존재하지 않는 사용자 또는 미션입니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"409": {
						"description": "이미 도전 중인 미션입니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 409,
											"message": "이미 도전 중인 미션입니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 500,
											"message": "서버 내부 오류",
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "미션 도전 API",
				"summary": "사용자가 특정 미션에 도전하는 엔드포인트입니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserMissionAddRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "HandleGetUserMissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserMissionListResponse_"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 사용자입니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionListApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 404,
											"message": "존재하지 않는 사용자입니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionListApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 500,
											"message": "서버 내부 오류",
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "진행 중인 미션 목록 조회 API",
				"summary": "사용자가 진행 중인 미션 목록을 조회하는 엔드포인트입니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지네이션 커서",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{userId}/missions/{missionId}": {
			"patch": {
				"operationId": "HandleCompleteUserMission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserMissionCompleteResponse_"
								}
							}
						}
					},
					"400": {
						"description": "진행 중인 미션이 아닙니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionCompleteApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 400,
											"message": "진행 중인 미션이 아닙니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 사용자 또는 미션입니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionCompleteApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 404,
											"message": "존재하지 않는 사용자 또는 미션입니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionCompleteApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 500,
											"message": "서버 내부 오류",
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "미션 완료 API",
				"summary": "진행 중인 미션을 완료로 변경하는 엔드포인트입니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "missionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/stores": {
			"post": {
				"operationId": "HandleAddStore",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_StoreAddResponse_"
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StoreAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 500,
											"message": "서버 내부 오류",
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "가게 추가 API",
				"summary": "특정 지역에 가게를 추가하는 엔드포인트입니다.",
				"tags": [
					"Stores"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StoreAddRequest"
							}
						}
					}
				}
			}
		},
...
		"/stores/{storeId}/missions": {
			"post": {
				"operationId": "HandleAddMission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_MissionAddResponse_"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 가게입니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MissionAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 404,
											"message": "존재하지 않는 가게입니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MissionAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 500,
											"message": "서버 내부 오류",
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "미션 추가 API",
				"summary": "특정 가게에 미션을 추가하는 엔드포인트입니다.",
				"tags": [
					"Missions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MissionAddRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "HandleGetStoreMissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_MissionListResponse_"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 가게입니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MissionListApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 404,
											"message": "존재하지 않는 가게입니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MissionListApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 500,
											"message": "서버 내부 오류",
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "가게 미션 목록 조회 API",
				"summary": "특정 가게의 미션 목록을 조회하는 엔드포인트입니다.",
				"tags": [
					"Missions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지네이션 커서",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}